Add Controller layers for all entities

+++ src/main/java/tn/esprit/projet_foyer/Repositories/FoyerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import tn.esprit.projet_foyer.Entities.Foyer;
 
 @Repository
-public interface FoyerRepository extends JpaRepository<Foyer, Integer> {
+public interface FoyerRepository extends JpaRepository<Foyer, Long> {
 }
